Rearganged tests
